// src/app/models/model.ts
export interface Task {
    readonly id?: number;
    readonly projectId?: number;
    readonly title: string;
    readonly description?: string;
    done: boolean;
    readonly order: number;
    readonly efforts?: TimeEfforts;
    readonly completed?: number;
    created: number;
}

export type TaskListFilterType = 'all' | 'open' | 'done';

export interface Project {
    readonly id?: number;
    readonly title: string;
    readonly description: string;
    readonly comments: Comment[];
}

export interface Tab {
    readonly id: any;
    readonly title: string;
}

export interface User {
    readonly id?: number;
    readonly name: string;
    readonly pictureUrl: string;
}

export interface Comment {
    readonly time: number;
    readonly user: User;
    readonly content: string;
}

export interface CommentUpdate {
    readonly index: number;
    readonly comment: Comment;
}

export type ActivityAlignment = 'left' | 'right';

export interface ActivitySliderSelection {
    start: number;
    end: number;
}

export interface ActivityBase {
    kind: string;
    id?: number;
    user: User;
    time: number;
    category: string;
    title: string;
    message: string;
}

export interface ProjectActivity extends ActivityBase {
    kind: 'project';
    projectId: number;
}

export type Activity = ProjectActivity;

export interface Tag {
    type: string;
    hashTag: string;
    title: string;
    link: string;
}

export interface InputPosition {
    top: number;
    left: number;
    caretOffset: number;
}

export interface HashTagInput {
    hashTag: string;
    position: InputPosition;
}

export interface TagSelection {
    tag: Tag;
    hashTagInput: HashTagInput;
}

export type DraggableType = 'task';

export interface TimeUnit {
    short: string;
    milliseconds: number;
}

export interface TimeEfforts {
    estimated: number;
    effective: number;
}

export interface ProjectSummary {
    project: Project;
    description: string;
    tasks: Task[];
    activities: ProjectActivity[];
}

export interface RasterizationData {
    time: number;
    weight: number;
}